be/cmd/objectives: moves the lying field type decls inside a in-service type package
