+++ backend/cmd/objectives/models/scalars/types.go
@@ -1,0 +1,6 @@
+package scalars
+
+type (
+	PlacementLength int
+	PlacementStart  int
+)
